pushed

+++ src/server/Object.cpp
@@ -31,16 +31,16 @@ void Object::move(const Direction dir)
 	switch (dir)
 	{
 	case Direction::SOUTH:
-		m_position.second+=1.0f;
+		m_position.second+=1;
 		break;
 	case Direction::EAST:
-		m_position.first+=1.0f;
+		m_position.first+=1;
 		break;
 	case Direction::NORTH:
-		m_position.second-=1.0f;
+		m_position.second-=1;
 		break;
 	case Direction::WEST:
-		m_position.first-=1.0f;
+		m_position.first-=1;
 		break;
 	default:
 		throw;
@@ -48,7 +48,7 @@ void Object::move(const Direction dir)
 	}
 }
 
-void Object::set_position(const std::pair<float, float> new_position)
+void Object::set_position(const std::pair<int, int> new_position)
 {
 	m_position = new_position;
 }
